requirements.txt
streamlit>=1.31.0
torch==2.1.2+cpu
numpy==1.26.4
transformers==4.36.2
google-genai
